Внес исправления в спринт 5
